refactor(libpod): move more content to libpod API
avoid the import to ../../main in preload
--> from '../../main/src/plugin/dockerode/libpod-dockerode';

related to https://github.com/podman-desktop/podman-desktop/issues/14361
