+++ packages/api/src/libpod/libpod.ts
@@ -39,3 +39,82 @@ export interface PlayKubeInfo {
   StopReport: { Err: string; Id: string }[];
   Volumes: { Name: string }[];
 }
+
+export interface ContainerCreateMountOption {
+  Name?: string;
+  Type: string;
+  Source: string;
+  Destination: string;
+  Driver?: string;
+  RW: boolean;
+  Propagation: string;
+  Options?: string[];
+}
+
+export interface ContainerCreatePortMappingOption {
+  container_port: number;
+  host_ip?: string;
+  host_port?: number;
+  protocol?: string;
+  range?: number;
+}
+
+export interface ContainerCreateNetNSOption {
+  nsmode: string;
+  value?: string;
+}
+
+export interface ContainerCreateHealthConfigOption {
+  Test?: string[];
+  Interval?: number;
+  Timeout?: number;
+  StartPeriod?: number;
+  Retries?: number;
+}
+
+export interface ContainerCreateNamedVolume {
+  Name: string;
+  Dest: string;
+  Options?: Array<string>;
+  IsAnonymous?: boolean;
+  SubPath?: string;
+}
+
+// represents a device request through the libPod API
+// only path is currently translated
+export interface PodmanDevice {
+  path: string;
+}
+
+export interface ContainerCreateOptions {
+  command?: string[];
+  entrypoint?: string | string[];
+  env?: { [key: string]: string };
+  pod?: string;
+  hostname?: string;
+  image?: string;
+  name?: string;
+  mounts?: Array<ContainerCreateMountOption>;
+  user?: string;
+  labels?: { [label: string]: string };
+  work_dir?: string;
+  portmappings?: Array<ContainerCreatePortMappingOption>;
+  stop_timeout?: number;
+  healthconfig?: ContainerCreateHealthConfigOption;
+  restart_policy?: string;
+  restart_tries?: number;
+  remove?: boolean;
+  seccomp_policy?: string;
+  seccomp_profile_path?: string;
+  cap_add?: Array<string>;
+  cap_drop?: Array<string>;
+  privileged?: boolean;
+  netns?: ContainerCreateNetNSOption;
+  read_only_filesystem?: boolean;
+  dns_server?: Array<Array<number>>;
+  hostadd?: Array<string>;
+  userns?: string;
+  volumes?: Array<ContainerCreateNamedVolume>;
+  selinux_opts?: string[];
+  devices?: PodmanDevice[];
+}
